github_tool/requirements.txt
certifi==2023.7.22
cffi==1.15.1
cryptography
distlib==0.3.7
docutils==0.20.1
filelock==3.12.4
github-backup==0.43.1
pipenv==2023.9.8
platformdirs==3.10.0
pycparser==2.21
Pygments
PyYAML
six==1.16.0
tabulate
virtualenv==20.24.5
PyGithub
GitPython
flake8==6.1.0